+++ libsolidity/analysis/TypeChecker.h
@@ -24,10 +24,10 @@
 
 #include <liblangutil/EVMVersion.h>
 
-#include <libsolidity/ast/Types.h>
 #include <libsolidity/ast/ASTAnnotations.h>
 #include <libsolidity/ast/ASTForward.h>
 #include <libsolidity/ast/ASTVisitor.h>
+#include <libsolidity/ast/Types.h>
 
 namespace langutil
 {
@@ -80,6 +80,8 @@ private:
 		FunctionCall const& _functionCall,
 		bool _abiEncoderV2
 	);
+
+	TypePointers typeCheckMetaTypeFunctionAndRetrieveReturnType(FunctionCall const& _functionCall);
 
 	/// Performs type checks and determines result types for type conversion FunctionCall nodes.
 	TypePointer typeCheckTypeConversionAndRetrieveReturnType(