+++ libsolidity/analysis/ViewPureChecker.cpp
@@ -16,14 +16,10 @@
 */
 
 #include <libsolidity/analysis/ViewPureChecker.h>
-
-#include <libevmasm/SemanticInformation.h>
-
 #include <libsolidity/ast/ExperimentalFeatures.h>
 #include <libyul/AsmData.h>
-
 #include <liblangutil/ErrorReporter.h>
-
+#include <libevmasm/SemanticInformation.h>
 #include <functional>
 
 using namespace std;
@@ -342,7 +338,9 @@ void ViewPureChecker::endVisit(MemberAccess const& _memberAccess)
 			{MagicType::Kind::ABI, "encodeWithSignature"},
 			{MagicType::Kind::Block, "blockhash"},
 			{MagicType::Kind::Message, "data"},
-			{MagicType::Kind::Message, "sig"}
+			{MagicType::Kind::Message, "sig"},
+			{MagicType::Kind::MetaType, "creationCode"},
+			{MagicType::Kind::MetaType, "runtimeCode"}
 		};
 		set<MagicMember> static const payableMembers{
 			{MagicType::Kind::Message, "value"}