+++ libyul/backends/evm/EVMDialect.cpp
@@ -38,13 +38,13 @@ using namespace dev::solidity;
 
 
 EVMDialect::EVMDialect(AsmFlavour _flavour, bool _objectAccess):
-	Dialect(_flavour), m_objectAccess(_objectAccess)
+	Dialect{_flavour}, m_objectAccess(_objectAccess)
 {
 	// The EVM instructions will be moved to builtins at some point.
 	if (!m_objectAccess)
 		return;
 
-	addFunction("datasize", 1, 1, true, [this](
+	addFunction("datasize", 1, 1, true, true, [this](
 		FunctionCall const& _call,
 		AbstractAssembly& _assembly,
 		std::function<void()>
@@ -58,7 +58,7 @@ EVMDialect::EVMDialect(AsmFlavour _flavour, bool _objectAccess):
 		else
 			_assembly.appendDataSize(m_subIDs.at(dataName));
 	});
-	addFunction("dataoffset", 1, 1, true, [this](
+	addFunction("dataoffset", 1, 1, true, true, [this](
 		FunctionCall const& _call,
 		AbstractAssembly& _assembly,
 		std::function<void()>
@@ -72,7 +72,7 @@ EVMDialect::EVMDialect(AsmFlavour _flavour, bool _objectAccess):
 		else
 			_assembly.appendDataOffset(m_subIDs.at(dataName));
 	});
-	addFunction("datacopy", 3, 0, false, [](
+	addFunction("datacopy", 3, 0, false, false, [](
 		FunctionCall const&,
 		AbstractAssembly& _assembly,
 		std::function<void()> _visitArguments
@@ -128,6 +128,7 @@ void EVMDialect::addFunction(
 	size_t _params,
 	size_t _returns,
 	bool _movable,
+	bool _literalArguments,
 	std::function<void(FunctionCall const&, AbstractAssembly&, std::function<void()>)> _generateCode
 )
 {
@@ -137,5 +138,6 @@ void EVMDialect::addFunction(
 	f.parameters.resize(_params);
 	f.returns.resize(_returns);
 	f.movable = _movable;
+	f.literalArguments = _literalArguments;
 	f.generateCode = std::move(_generateCode);
 }